possible fix for long files

+++ plugins/Ninjas/Sample.cpp
@@ -59,23 +59,21 @@ int Sample::loadSample ( std::string fp, std::vector <float> & samplevector, dou
 
 void Sample::createSlices ( Slice* slices, int n_slices )
 {
-    float sliceSize = ( float ) ( size*channels ) / ( float ) n_slices;
-
-    /*
+    long double sliceSize = ( long double ) ( size*channels ) / ( long double ) n_slices;
+    
     std::cout << "sliceSize :" << sliceSize
     << " x " << n_slices << " n_slices = "
     << n_slices * sliceSize << std::endl;
-    */
+    
 
     for ( int i = 0 ; i < n_slices; i++ )
     {
         slices[i].setSliceStart ( ( int ) i * sliceSize );
         slices[i].setSliceEnd ( ( ( int ) ( i+1 ) * sliceSize ) - 1 );
-        /*
-        std::cout << "Slice :" << i << " : "
+
+	std::cout << "Slice :" << i << " : "
         << slices[i].getSliceStart() << "->"
         << slices[i].getSliceEnd() << std::endl;
-        */
     }
 }
